+++ src/main/java/org/zalando/nakadi/service/subscription/state/ClosingState.java
@@ -5,7 +5,9 @@ import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.exceptions.NakadiRuntimeException;
 import org.zalando.nakadi.exceptions.runtime.MyNakadiRuntimeException1;
 import org.zalando.nakadi.service.subscription.model.Partition;
-import org.zalando.nakadi.service.subscription.zk.ZKSubscription;
+import org.zalando.nakadi.service.subscription.zk.ZkSubscription;
+import org.zalando.nakadi.service.subscription.zk.ZkSubscriptionClient;
+import org.zalando.nakadi.view.SubscriptionCursorWithoutToken;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -22,8 +24,8 @@ class ClosingState extends State {
     private final Supplier<Map<EventTypePartition, NakadiCursor>> uncommittedOffsetsSupplier;
     private final LongSupplier lastCommitSupplier;
     private Map<EventTypePartition, NakadiCursor> uncommittedOffsets;
-    private final Map<EventTypePartition, ZKSubscription> listeners = new HashMap<>();
-    private ZKSubscription topologyListener;
+    private final Map<EventTypePartition, ZkSubscription<SubscriptionCursorWithoutToken>> listeners = new HashMap<>();
+    private ZkSubscription<ZkSubscriptionClient.Topology> topologyListener;
 
     ClosingState(final Supplier<Map<EventTypePartition, NakadiCursor>> uncommittedOffsetsSupplier,
                  final LongSupplier lastCommitSupplier) {
@@ -41,7 +43,7 @@ class ClosingState extends State {
         } finally {
             if (null != topologyListener) {
                 try {
-                    topologyListener.cancel();
+                    topologyListener.close();
                 } finally {
                     topologyListener = null;
                 }
@@ -56,7 +58,12 @@ class ClosingState extends State {
         uncommittedOffsets = uncommittedOffsetsSupplier.get();
         if (!uncommittedOffsets.isEmpty() && timeToWaitMillis > 0) {
             scheduleTask(() -> switchState(new CleanupState()), timeToWaitMillis, TimeUnit.MILLISECONDS);
-            topologyListener = getZk().subscribeForTopologyChanges(() -> addTask(this::onTopologyChanged));
+            try {
+                topologyListener = getZk().subscribeForTopologyChanges(() -> addTask(this::onTopologyChanged));
+            } catch (final Exception e) {
+                switchState(new CleanupState(e));
+                return;
+            }
             reactOnTopologyChange();
         } else {
             switchState(new CleanupState());
@@ -68,19 +75,19 @@ class ClosingState extends State {
             throw new IllegalStateException(
                     "topologyListener should not be null when calling onTopologyChanged method");
         }
-        topologyListener.refresh();
         reactOnTopologyChange();
     }
 
-    private void reactOnTopologyChange() {
+    private void reactOnTopologyChange() throws NakadiRuntimeException {
+        final ZkSubscriptionClient.Topology topology = topologyListener.getData();
 
         // Collect current partitions state from Zk
         final Map<EventTypePartition, Partition> partitions = new HashMap<>();
-        getZk().runLocked(() -> Stream.of(getZk().listPartitions())
+        Stream.of(topology.getPartitions())
                 .filter(p -> getSessionId().equals(p.getSession()))
-                .forEach(p -> partitions.put(p.getKey(), p)));
+                .forEach(p -> partitions.put(p.getKey(), p));
+
         // Select which partitions need to be freed from this session
-        // Ithere
         final Set<EventTypePartition> freeRightNow = new HashSet<>();
         final Set<EventTypePartition> addListeners = new HashSet<>();
         for (final Partition p : partitions.values()) {
@@ -108,21 +115,17 @@ class ClosingState extends State {
         listeners.put(
                 key,
                 getZk().subscribeForOffsetChanges(
-                        key, () -> addTask(() -> this.offsetChanged(key))));
-        reactOnOffset(key);
-    }
-
-    private void offsetChanged(final EventTypePartition key) {
-        if (listeners.containsKey(key)) {
-            listeners.get(key).refresh();
-        }
+                        key, () -> addTask(() -> this.reactOnOffset(key))));
         reactOnOffset(key);
     }
 
     private void reactOnOffset(final EventTypePartition key) {
+        if (!listeners.containsKey(key)) {
+            return;
+        }
         final NakadiCursor newCursor;
         try {
-            newCursor = getContext().getCursorConverter().convert(key.getEventType(), getZk().getOffset(key));
+            newCursor = getContext().getCursorConverter().convert(key.getEventType(), listeners.get(key).getData());
         } catch (Exception ex) {
             throw new NakadiRuntimeException(ex);
         }
@@ -142,10 +145,10 @@ class ClosingState extends State {
         RuntimeException exceptionCaught = null;
         for (final EventTypePartition partitionKey : keys) {
             uncommittedOffsets.remove(partitionKey);
-            final ZKSubscription listener = listeners.remove(partitionKey);
+            final ZkSubscription<SubscriptionCursorWithoutToken> listener = listeners.remove(partitionKey);
             if (null != listener) {
                 try {
-                    listener.cancel();
+                    listener.close();
                 } catch (final RuntimeException ex) {
                     exceptionCaught = ex;
                     getLog().error("Failed to cancel offsets listener {}", listener, ex);