+++ src/main/java/org/zalando/nakadi/service/subscription/state/StreamingState.java
@@ -17,10 +17,12 @@ import org.zalando.nakadi.exceptions.ServiceUnavailableException;
 import org.zalando.nakadi.metrics.MetricUtils;
 import org.zalando.nakadi.repository.EventConsumer;
 import org.zalando.nakadi.service.subscription.model.Partition;
-import org.zalando.nakadi.service.subscription.zk.ZKSubscription;
+import org.zalando.nakadi.service.subscription.zk.ZkSubscription;
+import org.zalando.nakadi.service.subscription.zk.ZkSubscriptionClient;
 import org.zalando.nakadi.view.SubscriptionCursor;
 import org.zalando.nakadi.view.SubscriptionCursorWithoutToken;
 
+import java.io.Closeable;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -43,7 +45,7 @@ class StreamingState extends State {
     // The reasons for that if there are two partitions (p0, p1) and p0 is reassigned, if p1 is working
     // correctly, and p0 is not receiving any updates - reassignment won't complete.
     private final Map<EventTypePartition, Long> releasingPartitions = new HashMap<>();
-    private ZKSubscription topologyChangeSubscription;
+    private ZkSubscription<ZkSubscriptionClient.Topology> topologyChangeSubscription;
     private EventConsumer.ReassignableEventConsumer eventConsumer;
     private boolean pollPaused;
     private long committedEvents;
@@ -52,7 +54,7 @@ class StreamingState extends State {
     private Meter bytesSentMeter;
     // Uncommitted offsets are calculated right on exiting from Streaming state.
     private Map<EventTypePartition, NakadiCursor> uncommittedOffsets;
-    private ZKSubscription cursorResetSubscription;
+    private Closeable cursorResetSubscription;
 
     /**
      * Time that is used for commit timeout check. Commit timeout check is working only in case when there is something
@@ -74,7 +76,8 @@ class StreamingState extends State {
         this.eventConsumer = getContext().getTimelineService().createEventConsumer(null);
 
         // Subscribe for topology changes.
-        this.topologyChangeSubscription = getZk().subscribeForTopologyChanges(() -> addTask(this::topologyChanged));
+        this.topologyChangeSubscription =
+                getZk().subscribeForTopologyChanges(() -> addTask(this::reactOnTopologyChange));
         // and call directly
         reactOnTopologyChange();
         addTask(this::pollDataFromKafka);
@@ -261,7 +264,7 @@ class StreamingState extends State {
 
         if (null != topologyChangeSubscription) {
             try {
-                topologyChangeSubscription.cancel();
+                topologyChangeSubscription.close();
             } catch (final RuntimeException ex) {
                 getLog().warn("Failed to cancel topology subscription", ex);
             } finally {
@@ -280,25 +283,23 @@ class StreamingState extends State {
         }
 
         if (cursorResetSubscription != null) {
-            cursorResetSubscription.cancel();
+            try {
+                cursorResetSubscription.close();
+            } catch (IOException ignore) {
+            }
             cursorResetSubscription = null;
         }
     }
 
-    void topologyChanged() {
-        if (null != topologyChangeSubscription) {
-            topologyChangeSubscription.refresh();
+    void reactOnTopologyChange() {
+        if (null == topologyChangeSubscription) {
+            return;
         }
-        reactOnTopologyChange();
-    }
-
-    private void reactOnTopologyChange() {
-        getZk().runLocked(() -> {
-            final Partition[] assignedPartitions = Stream.of(getZk().listPartitions())
-                    .filter(p -> getSessionId().equals(p.getSession()))
-                    .toArray(Partition[]::new);
-            addTask(() -> refreshTopologyUnlocked(assignedPartitions));
-        });
+        final ZkSubscriptionClient.Topology topology = topologyChangeSubscription.getData();
+        final Partition[] assignedPartitions = Stream.of(topology.getPartitions())
+                .filter(p -> getSessionId().equals(p.getSession()))
+                .toArray(Partition[]::new);
+        addTask(() -> refreshTopologyUnlocked(assignedPartitions));
     }
 
     void refreshTopologyUnlocked(final Partition[] assignedPartitions) {
@@ -439,7 +440,7 @@ class StreamingState extends State {
     private void addToStreaming(final Partition partition) {
         final NakadiCursor cursor = createNakadiCursor(getZk().getOffset(partition.getKey()));
         getLog().info("Adding to streaming {} with start position {}", partition.getKey(), cursor);
-        final ZKSubscription subscription = getZk().subscribeForOffsetChanges(
+        final ZkSubscription<SubscriptionCursorWithoutToken> subscription = getZk().subscribeForOffsetChanges(
                 partition.getKey(),
                 () -> addTask(() -> offsetChanged(partition.getKey())));
         final PartitionData pd = new PartitionData(
@@ -467,16 +468,15 @@ class StreamingState extends State {
     void offsetChanged(final EventTypePartition key) {
         if (offsets.containsKey(key)) {
             final PartitionData data = offsets.get(key);
-            data.getSubscription().refresh();
 
-            final NakadiCursor cursor = createNakadiCursor(getZk().getOffset(key));
+            final NakadiCursor cursor = createNakadiCursor(data.getSubscription().getData());
 
             final PartitionData.CommitResult commitResult = data.onCommitOffset(cursor);
             if (commitResult.seekOnKafka) {
                 reconfigureKafkaConsumer(true);
             }
 
-	    if (commitResult.committedCount > 0) {
+            if (commitResult.committedCount > 0) {
                 committedEvents += commitResult.committedCount;
                 this.lastCommitMillis = System.currentTimeMillis();
                 streamToOutput();
@@ -503,7 +503,7 @@ class StreamingState extends State {
                     getLog().warn("Skipping commits: {}, commit={}, sent={}",
                             key, data.getCommitOffset(), data.getSentOffset());
                 }
-                data.getSubscription().cancel();
+                data.getSubscription().close();
             } catch (final RuntimeException ex) {
                 getLog().warn("Failed to cancel subscription, skipping exception", ex);
             }