+++ src/main/java/org/zalando/nakadi/service/subscription/zk/ZkSubscriptionClient.java
@@ -1,14 +1,18 @@
 package org.zalando.nakadi.service.subscription.zk;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.zalando.nakadi.domain.EventTypePartition;
 import org.zalando.nakadi.exceptions.NakadiRuntimeException;
 import org.zalando.nakadi.exceptions.ServiceUnavailableException;
+import org.zalando.nakadi.exceptions.runtime.MyNakadiRuntimeException1;
 import org.zalando.nakadi.exceptions.runtime.OperationTimeoutException;
 import org.zalando.nakadi.exceptions.runtime.ZookeeperException;
 import org.zalando.nakadi.service.subscription.model.Partition;
 import org.zalando.nakadi.service.subscription.model.Session;
 import org.zalando.nakadi.view.SubscriptionCursorWithoutToken;
 
+import java.io.Closeable;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
@@ -75,7 +79,7 @@ public interface ZkSubscriptionClient {
      *
      * @param listener method to call on any change of client list.
      */
-    ZKSubscription subscribeForSessionListChanges(Runnable listener);
+    ZkSubscription<List<String>> subscribeForSessionListChanges(Runnable listener) throws Exception;
 
     /**
      * Subscribe for topology changes.
@@ -83,9 +87,10 @@ public interface ZkSubscriptionClient {
      * @param listener called whenever /nakadi/subscriptions/{subscriptionId}/topology node is changed.
      * @return Subscription instance
      */
-    ZKSubscription subscribeForTopologyChanges(Runnable listener);
+    ZkSubscription<Topology> subscribeForTopologyChanges(Runnable listener) throws NakadiRuntimeException;
 
-    ZKSubscription subscribeForOffsetChanges(EventTypePartition key, Runnable commitListener);
+    ZkSubscription<SubscriptionCursorWithoutToken> subscribeForOffsetChanges(
+            EventTypePartition key, Runnable commitListener);
 
     /**
      * Returns current offset value for specified partition key. Offset includes timeline and version data.
@@ -130,9 +135,9 @@ public interface ZkSubscriptionClient {
      * Subscribes to cursor reset event.
      *
      * @param listener callback which is called when cursor reset happens
-     * @return {@link org.zalando.nakadi.service.subscription.zk.ZKSubscription}
+     * @return {@link Closeable}
      */
-    ZKSubscription subscribeForCursorsReset(Runnable listener)
+    Closeable subscribeForCursorsReset(Runnable listener)
             throws NakadiRuntimeException, UnsupportedOperationException;
 
     /**
@@ -152,4 +157,46 @@ public interface ZkSubscriptionClient {
      */
     void resetCursors(List<SubscriptionCursorWithoutToken> cursors, long timeout)
             throws OperationTimeoutException, ZookeeperException;
+
+    class Topology {
+        private final Partition[] partitions;
+        private final int version;
+
+        public Topology(
+                @JsonProperty("partitions") final Partition[] partitions,
+                @JsonProperty("version") final int version) {
+            this.partitions = partitions;
+            this.version = version;
+        }
+
+        public Partition[] getPartitions() {
+            return partitions;
+        }
+
+        public Topology withUpdatedPartitions(final Partition[] partitions) {
+            final Partition[] resultPartitions = Arrays.copyOf(this.partitions, this.partitions.length);
+            for (final Partition newValue : partitions) {
+                int selectedIdx = -1;
+                for (int idx = 0; idx < resultPartitions.length; ++idx) {
+                    if (resultPartitions[idx].getKey().equals(newValue.getKey())) {
+                        selectedIdx = idx;
+                    }
+                }
+                if (selectedIdx < 0) {
+                    throw new MyNakadiRuntimeException1(
+                            "Failed to find partition " + newValue.getKey() + " in " + this);
+                }
+                resultPartitions[selectedIdx] = newValue;
+            }
+            return new Topology(resultPartitions, version + 1);
+        }
+
+        @Override
+        public String toString() {
+            return "Topology{" +
+                    "partitions=" + Arrays.toString(partitions) +
+                    ", version=" + version +
+                    '}';
+        }
+    }
 }